SQL Script

+++ u25-Ejercicio3/src/main/java/com/team02/u25/ejercicio3/dao/IAlmacenDAO.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.team02.u25.ejercicio3.dto.Almacen;
 
 
-public interface IAlmacenDAO extends JpaRepository<Almacen, Long>{
+public interface IAlmacenDAO extends JpaRepository<Almacen, Integer>{
 
 }
